Rename + update with config file
